Cast to DOUBLE instead of DECIMAL

+++ models/aave/ethereum/aave_v2_ethereum_supply.sql
@@ -30,7 +30,7 @@ SELECT
     user AS depositor, 
     CAST(NULL AS VARCHAR(5)) as withdrawn_to,
     CAST(NULL AS VARCHAR(5)) AS liquidator,
-    CAST(amount AS DECIMAL(38,0)) AS amount,
+    CAST(amount AS DOUBLE) AS amount,
     evt_tx_hash,
     evt_index,
     evt_block_time,
