+++ models/blur/ethereum/blur_ethereum_events.sql
@@ -24,10 +24,10 @@ SELECT
     , CAST(json_extract_scalar(bm.sell, '$.tokenId') AS varchar) AS token_id
     , nft.standard AS token_standard
     , nft.name AS collection
-    , CASE WHEN CAST(JSON_EXTRACT_SCALAR(bm.buy, '$.amount') AS DECIMAL(38,0)) = 1 THEN 'Single Item Trade'
+    , CASE WHEN CAST(JSON_EXTRACT_SCALAR(bm.buy, '$.amount') AS DOUBLE) = 1 THEN 'Single Item Trade'
             ELSE 'Bundle Trade'
             END AS trade_type
-    , CAST(JSON_EXTRACT_SCALAR(bm.buy, '$.amount') AS DECIMAL(38,0)) AS number_of_items
+    , CAST(JSON_EXTRACT_SCALAR(bm.buy, '$.amount') AS DOUBLE) AS number_of_items
     , 'Trade' AS evt_type
     , CASE WHEN JSON_EXTRACT_SCALAR(bm.sell, '$.trader') = agg.contract_address THEN et."from" ELSE JSON_EXTRACT_SCALAR(bm.sell, '$.trader') END AS seller
     , CASE WHEN JSON_EXTRACT_SCALAR(bm.buy, '$.trader') = agg.contract_address THEN et."from" ELSE JSON_EXTRACT_SCALAR(bm.buy, '$.trader') END AS buyer
@@ -37,7 +37,7 @@ SELECT
         WHEN et."from" = JSON_EXTRACT_SCALAR(bm.sell, '$.trader') THEN 'Offer Accepted'
         ELSE 'Unknown'
         END AS trade_category
-    , CAST(JSON_EXTRACT_SCALAR(bm.buy, '$.price') AS DECIMAL(38,0)) AS amount_raw
+    , CAST(JSON_EXTRACT_SCALAR(bm.buy, '$.price') AS DOUBLE) AS amount_raw
     , CASE WHEN JSON_EXTRACT_SCALAR(bm.buy, '$.paymentToken') IN (0x0000000000000000000000000000000000000000, 0x0000000000a39bb272e79075ade125fd351887ac) THEN CAST(JSON_EXTRACT_SCALAR(bm.buy, '$.price') / POWER(10, 18) AS double)
         ELSE CAST(JSON_EXTRACT_SCALAR(bm.buy, '$.price') / POWER(10, pu.decimals) AS double)
         END AS amount_original
@@ -125,12 +125,12 @@ SELECT
     , nft_tok.standard AS token_standard
     , nft_tok.name AS collection
     , CASE WHEN json_extract_scalar(s.offer[0], '$.amount') = '1' THEN 'Single Item Trade' ELSE 'Bundle Trade' END AS trade_type
-    , CAST(json_extract_scalar(s.offer[0], '$.amount') AS DECIMAL(38,0)) AS number_of_items
+    , CAST(json_extract_scalar(s.offer[0], '$.amount') AS DOUBLE) AS number_of_items
     , 'Trade' AS evt_type
     , s.offerer AS seller
     , s.recipient AS buyer
     , 'Buy' AS trade_category
-    , CAST(json_extract_scalar(s.consideration[0], '$.amount') + json_extract_scalar(s.consideration[1], '$.amount') AS DECIMAL(38,0)) AS amount_raw
+    , CAST(json_extract_scalar(s.consideration[0], '$.amount') + json_extract_scalar(s.consideration[1], '$.amount') AS DOUBLE) AS amount_raw
     , CAST((json_extract_scalar(s.consideration[0], '$.amount')+json_extract_scalar(s.consideration[1], '$.amount'))/POWER(10, 18) AS double) AS amount_original
     , CAST(pu.price*(json_extract_scalar(s.consideration[0], '$.amount')+json_extract_scalar(s.consideration[1], '$.amount'))/POWER(10, 18) AS double) AS amount_usd
     , CASE WHEN json_extract_scalar(s.consideration[0], '$.token')=0x0000000000000000000000000000000000000000 THEN 'ETH' ELSE currency_tok.symbol END AS currency_symbol
@@ -150,7 +150,7 @@ SELECT
     , LEAST(CAST(json_extract_scalar(s.consideration[0], '$.amount') AS DOUBLE), CAST(json_extract_scalar(s.consideration[1], '$.amount') AS DOUBLE))/POWER(10, 18) AS royalty_fee_amount
     , pu.price*LEAST(CAST(json_extract_scalar(s.consideration[0], '$.amount') AS DOUBLE), CAST(json_extract_scalar(s.consideration[1], '$.amount') AS DOUBLE))/POWER(10, 18) AS royalty_fee_amount_usd
     , 100.0*LEAST(CAST(json_extract_scalar(s.consideration[0], '$.amount') AS DOUBLE), CAST(json_extract_scalar(s.consideration[1], '$.amount') AS DOUBLE))
-        /CAST(CAST(json_extract_scalar(s.consideration[0], '$.amount') AS DOUBLE)+CAST(json_extract_scalar(s.consideration[1], '$.amount') AS DOUBLE) AS DECIMAL(38,0)) AS royalty_fee_percentage
+        /CAST(CAST(json_extract_scalar(s.consideration[0], '$.amount') AS DOUBLE)+CAST(json_extract_scalar(s.consideration[1], '$.amount') AS DOUBLE) AS DOUBLE) AS royalty_fee_percentage
     , CASE WHEN json_extract_scalar(s.consideration[0], '$.token')=0x0000000000000000000000000000000000000000 THEN 'ETH' ELSE currency_tok.symbol END AS royalty_fee_currency_symbol
     , CASE WHEN json_extract_scalar(s.consideration[0], '$.recipient')!=s.recipient THEN json_extract_scalar(s.consideration[0], '$.recipient')
         ELSE json_extract_scalar(s.consideration[1], '$.recipient')
