+++ models/opensea/ethereum/opensea_ethereum_events.sql
@@ -16,13 +16,13 @@ FROM
                 amount_usd,
                 token_standard,
                 trade_type,
-                CAST(number_of_items AS DECIMAL(38,0)) number_of_items,
+                CAST(number_of_items AS DOUBLE) number_of_items,
                 trade_category,
                 evt_type,
                 seller,
                 buyer,
                 amount_original,
-                CAST(amount_raw AS DECIMAL(38,0)) amount_raw,
+                CAST(amount_raw AS DOUBLE) amount_raw,
                 currency_symbol,
                 currency_contract,
                 nft_contract_address,
@@ -57,13 +57,13 @@ FROM
                 ,case when trade_type <> 'Bundle Trade' and count(1) over (partition by tx_hash) > 1 then 'Bulk Purchase'
                       else trade_type
                  end as trade_type
-                ,CAST(number_of_items AS DECIMAL(38,0)) number_of_items
+                ,CAST(number_of_items AS DOUBLE) number_of_items
                 ,case when is_private then 'Private Sale' else trade_category end as trade_category -- Private sale can be purchasd by Buy/Offer accepted, but we surpress when it is Private sale here 
                 ,evt_type
                 ,seller
                 ,buyer
                 ,amount_original
-                ,CAST(amount_raw AS DECIMAL(38,0)) amount_raw
+                ,CAST(amount_raw AS DOUBLE) amount_raw
                 ,currency_symbol
                 ,currency_contract
                 ,nft_contract_address
@@ -98,13 +98,13 @@ FROM
                 ,case when trade_type <> 'Bundle Trade' and count(1) over (partition by tx_hash) > 1 then 'Bulk Purchase'
                       else trade_type
                  end as trade_type
-                ,CAST(number_of_items AS DECIMAL(38,0)) number_of_items
+                ,CAST(number_of_items AS DOUBLE) number_of_items
                 ,case when is_private then 'Private Sale' else trade_category end as trade_category -- Private sale can be purchasd by Buy/Offer accepted, but we surpress when it is Private sale here 
                 ,evt_type
                 ,seller
                 ,buyer
                 ,amount_original
-                ,CAST(amount_raw AS DECIMAL(38,0)) amount_raw
+                ,CAST(amount_raw AS DOUBLE) amount_raw
                 ,currency_symbol
                 ,currency_contract
                 ,nft_contract_address
